+++ src/main/java/com/strangegrotto/wealthdraft/networth/AssetProjector.java
@@ -1,5 +1,0 @@
-package com.strangegrotto.wealthdraft.networth;
-
-public interface AssetProjector<T extends AssetSnapshot> {
-    T getNextSnapshot(T currentSnapshot);
-}